requirements.txt
selenium
webdriver-manager
mysql-connector-python
pandas
plotly
seaborn
nbformat>=4.2.0
scipy
statsmodels
dash